camorama: use gtk4 compatible items
This is a patch I wrote a long time ago when gtk4 was still
in -rc stages. Rebase it on the top of upstream.

There are some visual differences between the previous and
the new approach, but it makes the logic closer to support Gtk4.

Keep the gtk4 UI file in sync with the changes, as much as
possibe (as we can't test it right now).

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>